% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_font_installed}
\alias{is_font_installed}
\title{Check if a font file with the specified name is
in the system's font directory}
\usage{
is_font_installed(font_name)
}
\arguments{
\item{font_name}{Name of font to be checked}
}
\value{
Boolean indicating font presence
}
\description{
Check if a font file with the specified name is
in the system's font directory
}
\examples{
is_font_installed("Comic Sans")
}
