% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifu_pal.R
\name{nifu_pal}
\alias{nifu_pal}
\title{Create a color palette in NIFU colours}
\usage{
nifu_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Type of palette}

\item{reverse}{Specify direction of colour palette}

\item{...}{Additional arguments to colorRampPalette}
}
\description{
Create a color palette in NIFU colours
}
