% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifu_pal.R
\name{scale_colour_nifu}
\alias{scale_colour_nifu}
\alias{scale_color_nifu}
\alias{scale_fill_nifu}
\title{NIFU-farger for omriss}
\usage{
scale_colour_nifu(palette = c("default"), alpha = 1, ...)

scale_color_nifu(palette = c("default"), alpha = 1, ...)

scale_fill_nifu(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Type palett}

\item{alpha}{Transparens, en verdi mellom 0 og 1}

\item{...}{Ekstra parametre for ggplot2::discrete_scale()}
}
\description{
NIFU-farger for omriss
}
