% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nifu.R
\name{theme_nifu}
\alias{theme_nifu}
\title{Plot theme for NIFU graphs}
\usage{
theme_nifu(
  base_size = 16,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ...
)
}
\arguments{
\item{base_size}{set base text size}

\item{base_family}{set base font family}

\item{base_line_size}{set base line size}

\item{base_rect_size}{set base rectangle size}

\item{...}{additional parameters passed to scale_*_nifu()}
}
\value{
styling for ggplot plot objects
}
\description{
Plot theme for NIFU graphs
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) + 
  geom_point() + 
  theme_nifu()

}
