% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifu_pal.R
\name{nifu_cols}
\alias{nifu_cols}
\title{Specify NIFU colours}
\usage{
nifu_cols(type = "main", ...)
}
\arguments{
\item{type}{Type of palette, defaulting to a discrete colour palette.
Possible types are c("main", "reds", "blues")}

\item{...}{Additional arguments passed to ggplot2}
}
\value{
vector of hex codes for use in colour and fill aesthetics
}
\description{
Specify NIFU colours
}
\examples{
nifu_cols()

}
