% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifu_pal.R
\name{nifu_pal}
\alias{nifu_pal}
\title{NIFU-palett for ggplot2-grafer}
\usage{
nifu_pal(palette = c("default"), alpha = 1)
}
\arguments{
\item{palette}{Type palett}

\item{alpha}{Transparens, en verdi mellom 0 og 1}
}
\description{
NIFU-palett for ggplot2-grafer
}
\examples{
library(scales)
show_col(nifu_pal("default")(6))
}
